# agent_docs/server-api-v1/openapi.json
{
  "components": {
    "securitySchemes": {
      "SeatBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Subjective Player Server",
    "version": "1.0.0-design"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/bootstrap": {
      "get": {
        "operationId": "get_bootstrap",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/BootstrapResponse"
                }
              }
            },
            "description": "Complete typed result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          }
        },
        "security": [
          {
            "SeatBearer": []
          }
        ]
      }
    },
    "/api/v1/content/{revision}": {
      "get": {
        "operationId": "get_revision",
        "parameters": [
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ContentResponse"
                }
              }
            },
            "description": "Complete typed result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          }
        },
        "security": [
          {
            "SeatBearer": []
          }
        ]
      }
    },
    "/api/v1/games/{game_id}/ack": {
      "post": {
        "operationId": "post_ack",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Game-Epoch",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Audience-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Attachment-Epoch",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./player-api-v1.schema.json#/$defs/AckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/AckResponse"
                }
              }
            },
            "description": "Complete typed result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          }
        },
        "security": [
          {
            "SeatBearer": []
          }
        ]
      }
    },
    "/api/v1/games/{game_id}/attachment": {
      "post": {
        "operationId": "post_attachment",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Game-Epoch",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Audience-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./player-api-v1.schema.json#/$defs/AttachmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/AttachmentResponse"
                }
              }
            },
            "description": "Complete typed result"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "410": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./player-api-v1.schema.json#/$defs/ApiError"
                }
              }
            },
            "description": "Typed refusal or unavailable state; see finite error union"
          }
        },
        "security": [
          {
            "SeatBearer": []
          }
        ]
      }
    },
...
  },
...
}